Added Assignment 1
Assignment: Find a source text and manually perform one of the mashup
techniques below (or one of your own invention.) Create a webpage with
the results using some combination of HTML, CSS, and/or JavaScript.
Host the page on github pages (or your own server). There is no need
for programming for this assignment, it's just about getting set up in
an environment and starting to think about creative ways to play with
text. However, you may choose to include animated or interactive
elements if you like. Think about creative ways for the page to be
"self-documenting", i.e. instructions for the text mashup, references,
etc.